src/zoom/pyproject.toml
[tool.poetry]
name = "zoom-extractor"
version = "0.1.0"
description = ""
authors = ["Ed-Fi Alliance, LLC, and contributors"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
zoomus = "^1.1.3"
pandas = "^1.1.1"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
black = "^20.8b1"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
